Init commit.
The project template is a simplfied version of https://github.com/twoscoops/django-twoscoops-project.
I merge multiple requirements and settings files into a single file because this project is designed for simple tests and demos.

The dolphin app contains all the models. These models are based on the famous Sakila Sample Database.